Simplifies configuration.
