+++ jest.config.js
@@ -1,12 +1,4 @@
 module.exports = {
     coverageDirectory: '.coverage',
-    coverageThreshold: {
-        global: {
-            branches: 100,
-            functions: 100,
-            lines: 100,
-            statements: 100,
-        },
-    },
     testEnvironment: 'node',
 };
