+++ package.json
@@ -68,7 +68,6 @@
         "@typescript-eslint/eslint-plugin": "5.9.1",
         "@typescript-eslint/parser": "5.9.1",
         "confusing-browser-globals": "1.0.11",
-        "dot-prop": "6.0.1",
         "eslint-config-prettier": "8.3.0",
         "eslint-plugin-eslint-comments": "3.2.0",
         "eslint-plugin-get-off-my-lawn": "3.0.0",
@@ -85,6 +84,7 @@
         "eslint-plugin-security": "1.4.0",
         "eslint-plugin-unicorn": "40.0.0",
         "fs-extra": "10.0.0",
+        "merge-anything": "4.0.2",
         "semver": "7.3.5",
         "typescript": "4.5.4"
     },
