+++ src/comments.js
@@ -1,0 +1,10 @@
+module.exports = {
+    plugins: ['eslint-comments'],
+    rules: {
+        'eslint-comments/disable-enable-pair': 'error',
+        'eslint-comments/no-duplicate-disable': 'error',
+        'eslint-comments/no-unlimited-disable': 'error',
+        'eslint-comments/no-unused-disable': 'error',
+        'eslint-comments/no-unused-enable': 'error',
+    },
+};
