--- src/eslint-comments.js
@@ -1,7 +1,0 @@
-module.exports = {
-    'eslint-comments/disable-enable-pair': 'error',
-    'eslint-comments/no-duplicate-disable': 'error',
-    'eslint-comments/no-unlimited-disable': 'error',
-    'eslint-comments/no-unused-disable': 'error',
-    'eslint-comments/no-unused-enable': 'error',
-};
