--- src/jest-formatting.js
@@ -1,9 +1,0 @@
-module.exports = {
-    'jest-formatting/padding-around-after-all-blocks': 'error',
-    'jest-formatting/padding-around-after-each-blocks': 'error',
-    'jest-formatting/padding-around-before-all-blocks': 'error',
-    'jest-formatting/padding-around-before-each-blocks': 'error',
-    'jest-formatting/padding-around-describe-blocks': 'error',
-    'jest-formatting/padding-around-expect-groups': 'error',
-    'jest-formatting/padding-around-test-blocks': 'error',
-};
