+++ src/objects.js
@@ -1,5 +1,8 @@
 module.exports = {
-    'objects/no-object-properties-first-line': 'error',
-    'objects/no-object-properties-last-line': 'error',
-    'objects/no-object-properties-one-line': 'error',
+    plugins: ['objects'],
+    rules: {
+        'objects/no-object-properties-first-line': 'error',
+        'objects/no-object-properties-last-line': 'error',
+        'objects/no-object-properties-one-line': 'error',
+    },
 };
