--- src/react-hooks.js
@@ -1,4 +1,0 @@
-module.exports = {
-    'react-hooks/exhaustive-deps': 'error',
-    'react-hooks/rules-of-hooks': 'error',
-};
