+++ src/security.js
@@ -1,10 +1,13 @@
 module.exports = {
-    'security/detect-buffer-noassert': 'error',
-    'security/detect-child-process': 'error',
-    'security/detect-disable-mustache-escape': 'error',
-    'security/detect-eval-with-expression': 'error',
-    'security/detect-no-csrf-before-method-override': 'error',
-    'security/detect-possible-timing-attacks': 'error',
-    'security/detect-pseudoRandomBytes': 'error',
-    'security/detect-unsafe-regex': 'error',
+    plugins: ['security'],
+    rules: {
+        'security/detect-buffer-noassert': 'error',
+        'security/detect-child-process': 'error',
+        'security/detect-disable-mustache-escape': 'error',
+        'security/detect-eval-with-expression': 'error',
+        'security/detect-no-csrf-before-method-override': 'error',
+        'security/detect-possible-timing-attacks': 'error',
+        'security/detect-pseudoRandomBytes': 'error',
+        'security/detect-unsafe-regex': 'error',
+    },
 };
